add dots handlers
